move work components
