+++ app/components/work/work-left.tsx
@@ -1,4 +1,4 @@
-import { WorkProps } from './WorkRight';
+import { WorkProps } from './work-right';
 
 export function WorkLeft({ children, progress }: WorkProps) {
   let translateY = Math.max(0, 50 - progress * 3 * 50);
