+++ components/Providers/ThemeProvider.tsx
@@ -1,7 +1,6 @@
 'use client';
 
-import { ThemeProvider as NextThemesProvider } from 'next-themes';
-import type { ThemeProviderProps } from 'next-themes/dist/types';
+import { ThemeProvider as NextThemesProvider, ThemeProviderProps } from 'next-themes';
 
 export default function ThemeProvider({ children, ...props }: ThemeProviderProps) {
   return <NextThemesProvider {...props}>{children}</NextThemesProvider>;
